+++ dc-cudami-server/dc-cudami-server-backend-api/src/main/java/de/digitalcollections/cudami/server/backend/api/repository/identifiable/entity/work/ItemRepository.java
@@ -4,6 +4,8 @@ import de.digitalcollections.cudami.server.backend.api.repository.identifiable.e
 import de.digitalcollections.model.identifiable.entity.DigitalObject;
 import de.digitalcollections.model.identifiable.entity.work.Item;
 import de.digitalcollections.model.identifiable.entity.work.Work;
+import de.digitalcollections.model.list.paging.PageRequest;
+import de.digitalcollections.model.list.paging.PageResponse;
 import java.util.Set;
 import java.util.UUID;
 
@@ -12,7 +14,7 @@ public interface ItemRepository extends EntityRepository<Item> {
 
   boolean addWork(UUID itemUuid, UUID workUuid);
 
-  Set<DigitalObject> getDigitalObjects(UUID itemUuid);
+  PageResponse<DigitalObject> findDigitalObjects(UUID itemUuid, PageRequest pageRequest);
 
   Set<Work> getWorks(UUID itemUuid);
 }