+++ dc-cudami-server/dc-cudami-server-webapp/src/main/java/de/digitalcollections/cudami/server/controller/identifiable/entity/work/ItemController.java
@@ -10,8 +10,10 @@ import de.digitalcollections.model.identifiable.entity.DigitalObject;
 import de.digitalcollections.model.identifiable.entity.work.Item;
 import de.digitalcollections.model.identifiable.entity.work.Work;
 import de.digitalcollections.model.list.filtering.FilterCriterion;
+import de.digitalcollections.model.list.paging.PageRequest;
 import de.digitalcollections.model.list.paging.PageResponse;
 import de.digitalcollections.model.list.sorting.Order;
+import de.digitalcollections.model.list.sorting.Sorting;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -176,17 +178,23 @@ public class ItemController extends AbstractIdentifiableController<Item> {
     return new ResponseEntity<>(result, result != null ? HttpStatus.OK : HttpStatus.NOT_FOUND);
   }
 
-  @Operation(summary = "Get digital objects of this item")
+  @Operation(summary = "Get paged list of digital objects of this item")
   @GetMapping(
       value = {
         "/v6/items/{uuid}/digitalobjects",
-        "/v2/items/{uuid}/digitalobjects",
-        "/latest/items/{uuid}/digitalobjects"
       },
       produces = MediaType.APPLICATION_JSON_VALUE)
-  public Set<DigitalObject> getDigitalObjects(
-      @Parameter(name = "uuid", description = "UUID of the item") @PathVariable UUID uuid) {
-    return itemService.getDigitalObjects(uuid);
+  public PageResponse<DigitalObject> getDigitalObjects(
+      @Parameter(name = "uuid", description = "UUID of the item") @PathVariable UUID uuid,
+      @RequestParam(name = "pageNumber", required = false, defaultValue = "0") int pageNumber,
+      @RequestParam(name = "pageSize", required = false, defaultValue = "5") int pageSize,
+      @RequestParam(name = "sortBy", required = false) List<Order> sortBy) {
+    PageRequest pageRequest = new PageRequest(null, pageNumber, pageSize);
+    if (sortBy != null) {
+      Sorting sorting = new Sorting(sortBy);
+      pageRequest.setSorting(sorting);
+    }
+    return itemService.findDigitalObjects(uuid, pageRequest);
   }
 
   @Operation(summary = "Get works embodied in an item")