+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 - find by `searchTerm` looks for the existence of the single words;
   to look for the words as phrase, embed them in double quotes (but it will be much slower)
+- retrieval of DigitalObjects for an Item now works with paging
 
 ### Removed